Moving buttons back to Graph
