+++ src/components/pages/SearchResultsPage.js
@@ -7,7 +7,6 @@ import StockTable from "../singular/StockTable";
 
 const SearchResultsPage = () => {
   const location = useLocation();
-  const [selectedFilter, setSelectedFilter] = useState("Today");
 
   // Currently hardcoded but will eventually come from API
   const [chartData, setChartData] = useState([
@@ -26,13 +25,6 @@ const SearchResultsPage = () => {
     "Sep",
   ]);
 
-  const filters = {
-    DAY: "Today",
-    FIVE: "Past 5 days",
-    MONTH: "Past month",
-    YEAR: "Past Year",
-  };
-
   const [stockData, setStockData] = useState([
     {
       symbol: "AAPL",
@@ -46,19 +38,13 @@ const SearchResultsPage = () => {
     },
   ]);
 
-  const handleFilterSelection = (filter) => {
-    getData(filter);
-    setSelectedFilter(filter);
-    // logic to update chart data based on selected filter
-  };
-
   /*
   Callback used to get more data based on the filter. Each time any of the buttons 
   are clicked this will be called to get more data. This will update the chart data which 
   will then re-render the graph
   */
   const getData = (filter) => {
-    if (filter === "year") {
+    if (filter === "Past Year") {
       setChartData([]);
     } else {
       setChartData(chartData.concat(chartData));
@@ -93,10 +79,6 @@ const SearchResultsPage = () => {
               )}{" "}
               ({stockData[0].percentChanged}
               %)
-              <p className="inline text-light-gray font-light">
-                {" "}
-                {selectedFilter}
-              </p>
             </p>
           </div>
           <div className="w-11/12 h-4/5 mx-auto my-10">
@@ -105,42 +87,6 @@ const SearchResultsPage = () => {
               categories={categories}
               getData={getData}
             />
-            <div className="flex mt-4 justify-center">
-              <button
-                className={`py-2 px-4 text-white font-semibold border-b-2 border-dark-gray hover:bg-another-gray ${
-                  selectedFilter === filters.DAY ? "border-b-green active" : ""
-                }`}
-                onClick={() => handleFilterSelection(filters.DAY)}
-              >
-                D
-              </button>
-              <button
-                className={`py-2 px-4 text-white font-semibold border-b-2 border-dark-gray hover:bg-another-gray ${
-                  selectedFilter === filters.FIVE ? "border-b-green active" : ""
-                }`}
-                onClick={() => handleFilterSelection(filters.FIVE)}
-              >
-                5D
-              </button>
-              <button
-                className={`py-2 px-4 text-white font-semibold border-b-2 border-dark-gray hover:bg-another-gray ${
-                  selectedFilter === filters.MONTH
-                    ? "border-b-green active"
-                    : ""
-                }`}
-                onClick={() => handleFilterSelection(filters.MONTH)}
-              >
-                M
-              </button>
-              <button
-                className={`py-2 px-4 text-white font-semibold border-b-2 border-dark-gray hover:bg-another-gray ${
-                  selectedFilter === filters.YEAR ? "border-b-green active" : ""
-                }`}
-                onClick={() => handleFilterSelection(filters.YEAR)}
-              >
-                Y
-              </button>
-            </div>
             <StockTable stockData={stockData} getData={getData} />
           </div>
         </div>
